SimpleCoroutineAwaiter: prohibit `WaitForSeconds` in Edit Mode
Prohibit use of `WaitForSeconds` and `WaitForSecondsRealtime` in coroutines
that are running in Edit Mode.

`WaitForSeconds` will return immediately when used in Edit Mode, because the
waiting logic is implemented inside the Unity game loop.

`WaitForSecondsRealtime` will waits for an unpredictable time period in Edit
Mode because it relies on `Time.realtimeSinceStartup`, which isn't correctly
updated in Edit Mode.