AsyncUtilTests: split into PlayModeTests and EditModeTests
Added a new editor window for running tests in Edit Mode. Similar to the
existing tests in Play Mode, the Edit Mode tests are triggered by clicking
buttons in a GUI. The new Edit Mode Tests window can be opened from the Unity
menu by going to: Window => AsyncAwaitUtil => "Edit Mode Tests".

The test methods are shared between the Play Mode GUI and Edit Mode GUI and are
located in `AsyncUtilTests.cs` (as before).

Note 1: In order for all of the Edit Mode tests to run successfully, the Unity
version must be 2018.2.0 or higher! Earlier versions of Unity have fundamental
issues regarding usage of async/await in Editor scripts. See the following Unity
Forum thread for details:

https://forum.unity.com/threads/async-await-in-editor-script.481276/

Note 2: For maintenance purposes, it would be better if the Edit Mode tests were
implemented as normal unit tests that could be run by the Unity Test Runner. I
attempted to do this in Unity 2017.1.1f1, but found it impossible due to poor
async/await support in that version of Unity. However, it may be possible in
newer versions of Unity (I haven't tried).

EditModeTests: fix compile error